requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.6
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
